refactor: nuke unused cards, used cards to .tsx
