+++ components/index/cards/button.tsx
@@ -1,7 +1,20 @@
-import React from 'react'
+/** @jsxImportSource theme-ui */
 import { Box, Button, Text } from 'theme-ui'
 import ReactTooltip from '../../react-tooltip'
 import Icon from '@hackclub/icons'
+
+type ButtonsProps = {
+  children: React.ReactNode
+  icon?: string
+  customIcon?: React.ReactNode
+  id: string
+  content?: React.ReactNode
+  link?: string
+  primary?: boolean | string
+  overrideColor?: string
+  zIndex?: number
+  sx?: any
+}
 
 export default function Buttons({
   children,
@@ -13,14 +26,15 @@ export default function Buttons({
   primary,
   overrideColor,
   zIndex,
+  sx,
   ...props
-}) {
+}: ButtonsProps) {
   let fontWeight = primary ? '700' : '400'
 
   return (
     <Box
       as="button"
-      sx={{ background: 'transparent', border: 'none', color: 'white', zIndex: zIndex ||0 }}
+      sx={{ background: 'transparent', border: 'none', color: 'white', zIndex: zIndex ||0, ...sx }}
       py={1}
       tabIndex={-1}
     >
@@ -30,7 +44,7 @@ export default function Buttons({
         data-effect="solid"
         data-tip
         sx={{
-          background: primary || overrideColor || 'rgb(255, 255, 255, 0.3)',
+          background: (typeof primary === 'string' ? primary : undefined) || overrideColor || 'rgb(255, 255, 255, 0.3)',
           borderRadius: '100px',
           border: 'none',
           display: 'flex',
@@ -40,7 +54,6 @@ export default function Buttons({
           py: primary ? '12px' : 2,
           width: 'fit-content',
           textTransform: 'none',
-          fontWeight: '400',
           fontSize: primary ? ['18px', '20px', '22px'] : [1, '16px', '18px'],
           backdropFilter: 'blur(2px)',
           fontWeight: fontWeight,
@@ -58,10 +71,9 @@ export default function Buttons({
           </Box>
         ) : (
           <Icon
-            glyph={icon || 'plus-fill'}
-            sx={{ color: 'inherit', marginRight: 2 }}
+            glyph={(icon || 'plus-fill') as any}
             size={24}
-            mr={2}
+            style={{ color: 'inherit', marginRight: 2 }}
           />
         )}
         <Text sx={{ fontFamily: 'Phantom Sans', textAlign: 'left' }}>
