+++ components/index/cards/card-model.tsx
@@ -1,11 +1,9 @@
 import Icon from '../../icon'
 import { Box, Card, Flex, Image, Link, Text } from 'theme-ui'
-import { Zoom } from 'react-reveal'
 import ReactTooltip from '../../react-tooltip'
 import Comma from '../../comma'
 
 /** @jsxImportSource theme-ui */
-
 const CardModel = ({
   background,
   children,
@@ -23,6 +21,23 @@ const CardModel = ({
   filter,
   visible=false,
   ...props
+}: {
+  [x: string]: any
+  background?: any
+  children?: any
+  image?: any
+  image_fit?: any
+  link?: any
+  highlight?: any
+  github_link?: any
+  badge?: any
+  text?: any
+  color?: any
+  stars?: any
+  delay?: any
+  position?: any
+  filter?: any
+  visible?: boolean
 }) => (
   // <Zoom delay={delay}>
   <Card
