+++ components/index/cards/flavortown.tsx
@@ -1,8 +1,7 @@
+/** @jsxImportSource theme-ui */
 import CardModel from './card-model'
 import { Box, Flex, Grid, Image, Text } from 'theme-ui'
 import Buttons from './button'
-
-/** @jsxImportSource theme-ui */
 
 export default function Flavortown() {
   return (
