+++ components/index/cards/hackathons.tsx
@@ -1,10 +1,10 @@
+/** @jsxImportSource theme-ui */
 import CardModel from './card-model'
 import { Box, Flex, Grid, Image, Link, Text } from 'theme-ui'
 import Buttons from './button'
 import Dot from '../../dot'
 import { formatDate } from '../../../lib/dates'
 
-/** @jsxImportSource theme-ui */
 const Cover = () => (
   <Box
     sx={{
