+++ components/index/cards/hcb.tsx
@@ -1,9 +1,7 @@
+/** @jsxImportSource theme-ui */
 import CardModel from './card-model'
 import { Box, Grid, Heading, Text } from 'theme-ui'
-import { Fade } from 'react-reveal'
 import Buttons from './button'
-
-/** @jsxImportSource theme-ui */
 
 export default function Bank({ data }) {
   return (
@@ -11,7 +9,6 @@ export default function Bank({ data }) {
       <CardModel
         color="white"
         sx={{
-          backgroundColor: 'dark',
           minHeight: ['300px', '400px', '380px'],
           backgroundColor: 'darkless',
           backgroundImage: `linear-gradient(to bottom,rgba(0, 0, 0, 0.8), rgba(0, 0, 0, 0.6)), url('/home/hackathons-bg.webp')`,
