+++ components/index/cards/mailing-list.tsx
