+++ components/index/cards/sprig-console.tsx
@@ -1,9 +1,8 @@
+/** @jsxImportSource theme-ui */
 import { Box, Grid, Image, Text } from 'theme-ui'
 import Buttons from './button'
 import CardModel from './card-model'
 import Tilt from '../../tilt'
-
-/** @jsxImportSource theme-ui */
 
 export default function SprigConsole({ stars, consoleCount }) {
   return (
